requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.5
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-admin-interface==0.16.3
django-colorfield==0.4.2
django-crispy-forms==1.12.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-froala-editor==3.2.2
django-heroku==0.3.1
django-material-admin==1.8.5
django-widget-tweaks==1.4.8
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
Pillow==8.3.0
platformdirs==2.2.0
psycopg2==2.9.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.7.2
virtualenvwrapper-win==1.2.6
whitenoise==5.3.0
